requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
scikit-image==0.21.0
matplotlib==3.7.2
requests==2.31.0
geopy==2.3.0
Pillow-SIMD==9.0.0
tensorflow==2.13.0
